+++ src/tasks/cascadiacode.rs
@@ -1,0 +1,26 @@
+use crate::lib::{
+    ghrafont::GhraFont,
+    task::{self, Status, Task},
+};
+
+pub fn task() -> Task {
+    Task {
+        name: String::from("cascadiacode"),
+        sync,
+        update,
+    }
+}
+
+const GHRA_FONT: GhraFont = GhraFont {
+    asset_re: r"^Cascadia.ttf$",
+    font_suffix: ".ttf",
+    repo: ("microsoft", "cascadia-code"),
+};
+
+fn sync() -> task::Result {
+    GHRA_FONT.sync()
+}
+
+fn update(sync: Status) -> task::Result {
+    GHRA_FONT.update(sync)
+}