+++ src/utils/github.rs
@@ -21,6 +21,16 @@ pub struct Asset {
     state: String,
     updated_at: String,
 }
+impl Asset {
+    pub fn new() -> Self {
+        Self {
+            browser_download_url: String::new(),
+            name: String::new(),
+            state: String::new(),
+            updated_at: String::new(),
+        }
+    }
+}
 
 #[derive(Clone, Debug, Deserialize)]
 pub struct Release {